migration doesn't try to create tables with long table names when table_alias_length > 30 -- this is for better comaptibility with Oracle. (see http://code.google.com/p/rubycas-server/issues/detail?id=15)
git-svn-id: https://rubycas-server.googlecode.com/svn/trunk@213 fffcb96a-a727-0410-ad3e-7b35c796b8d7

+++ lib/casserver/models.rb
@@ -75,45 +75,51 @@ module CASServer::Models
 
   class CreateCASServer < V 0.1
     def self.up
-      $LOG.info("Migrating database")
+      if ActiveRecord::Base.connection.table_alias_length > 30
+        $LOG.info("Creating database with long table names...")
+        
+        create_table :casserver_login_tickets, :force => true do |t|
+          t.column :ticket,     :string,   :null => false
+          t.column :created_on, :timestamp, :null => false
+          t.column :consumed,   :datetime, :null => true
+          t.column :client_hostname, :string, :null => false
+        end
       
-      create_table :casserver_login_tickets, :force => true do |t|
-        t.column :ticket,     :string,   :null => false
-        t.column :created_on, :timestamp, :null => false
-        t.column :consumed,   :datetime, :null => true
-        t.column :client_hostname, :string, :null => false
-      end
-    
-      create_table :casserver_service_tickets, :force => true do |t|
-        t.column :ticket,     :string,    :null => false
-        t.column :service,    :string,    :null => false
-        t.column :created_on, :timestamp, :null => false
-        t.column :consumed,   :datetime, :null => true
-        t.column :client_hostname, :string, :null => false
-        t.column :username,   :string,  :null => false
-        t.column :type,       :string,   :null => false
-        t.column :proxy_granting_ticket_id, :integer, :null => true
-      end
-      
-      create_table :casserver_ticket_granting_tickets, :force => true do |t|
-        t.column :ticket,     :string,    :null => false
-        t.column :created_on, :timestamp, :null => false
-        t.column :client_hostname, :string, :null => false
-        t.column :username,   :string,    :null => false
-      end
-      
-      create_table :casserver_proxy_granting_tickets, :force => true do |t|
-        t.column :ticket,     :string,    :null => false
-        t.column :created_on, :timestamp, :null => false
-        t.column :client_hostname, :string, :null => false
-        t.column :iou,        :string,    :null => false
-        t.column :service_ticket_id, :integer, :null => false
+        create_table :casserver_service_tickets, :force => true do |t|
+          t.column :ticket,     :string,    :null => false
+          t.column :service,    :string,    :null => false
+          t.column :created_on, :timestamp, :null => false
+          t.column :consumed,   :datetime, :null => true
+          t.column :client_hostname, :string, :null => false
+          t.column :username,   :string,  :null => false
+          t.column :type,       :string,   :null => false
+          t.column :proxy_granting_ticket_id, :integer, :null => true
+        end
+        
+        create_table :casserver_ticket_granting_tickets, :force => true do |t|
+          t.column :ticket,     :string,    :null => false
+          t.column :created_on, :timestamp, :null => false
+          t.column :client_hostname, :string, :null => false
+          t.column :username,   :string,    :null => false
+        end
+        
+        create_table :casserver_proxy_granting_tickets, :force => true do |t|
+          t.column :ticket,     :string,    :null => false
+          t.column :created_on, :timestamp, :null => false
+          t.column :client_hostname, :string, :null => false
+          t.column :iou,        :string,    :null => false
+          t.column :service_ticket_id, :integer, :null => false
+        end
       end
     end
     
     def self.down
-      drop_table :casserver_service_tickets
-      drop_table :casserver_login_tickets
+      if ActiveRecord::Base.connection.table_alias_length > 30
+        drop_table :casserver_proxy_granting_tickets
+        drop_table :casserver_ticket_granting_tickets
+        drop_table :casserver_service_tickets
+        drop_table :casserver_login_tickets
+      end
     end
   end
   
@@ -121,18 +127,60 @@ module CASServer::Models
   # See http://code.google.com/p/rubycas-server/issues/detail?id=15
   class ShortenTableNames < V 0.5
     def self.up
-      $LOG.info("Shortening table names")
-      rename_table :casserver_login_tickets, :casserver_lt
-      rename_table :casserver_service_tickets, :casserver_st
-      rename_table :casserver_ticket_granting_tickets, :casserver_tgt
-      rename_table :casserver_proxy_granting_tickets, :casserver_pgt
+      if ActiveRecord::Base.connection.table_alias_length > 30
+        $LOG.info("Shortening table names")
+        rename_table :casserver_login_tickets, :casserver_lt
+        rename_table :casserver_service_tickets, :casserver_st
+        rename_table :casserver_ticket_granting_tickets, :casserver_tgt
+        rename_table :casserver_proxy_granting_tickets, :casserver_pgt
+      else
+        create_table :casserver_lt, :force => true do |t|
+          t.column :ticket,     :string,   :null => false
+          t.column :created_on, :timestamp, :null => false
+          t.column :consumed,   :datetime, :null => true
+          t.column :client_hostname, :string, :null => false
+        end
+      
+        create_table :casserver_st, :force => true do |t|
+          t.column :ticket,     :string,    :null => false
+          t.column :service,    :string,    :null => false
+          t.column :created_on, :timestamp, :null => false
+          t.column :consumed,   :datetime, :null => true
+          t.column :client_hostname, :string, :null => false
+          t.column :username,   :string,  :null => false
+          t.column :type,       :string,   :null => false
+          t.column :proxy_granting_ticket_id, :integer, :null => true
+        end
+        
+        create_table :casserver_tgt, :force => true do |t|
+          t.column :ticket,     :string,    :null => false
+          t.column :created_on, :timestamp, :null => false
+          t.column :client_hostname, :string, :null => false
+          t.column :username,   :string,    :null => false
+        end
+        
+        create_table :casserver_pgt, :force => true do |t|
+          t.column :ticket,     :string,    :null => false
+          t.column :created_on, :timestamp, :null => false
+          t.column :client_hostname, :string, :null => false
+          t.column :iou,        :string,    :null => false
+          t.column :service_ticket_id, :integer, :null => false
+        end
+      end
     end
     
     def self.down
-      rename_table :casserver_lt, :cassserver_login_tickets
-      rename_table :casserver_st, :casserver_service_tickets
-      rename_table :casserver_tgt, :casserver_ticket_granting_tickets
-      rename_table :casserver_pgt, :casserver_proxy_granting_tickets
+      if ActiveRecord::Base.connection.table_alias_length > 30
+        rename_table :casserver_lt, :cassserver_login_tickets
+        rename_table :casserver_st, :casserver_service_tickets
+        rename_table :casserver_tgt, :casserver_ticket_granting_tickets
+        rename_table :casserver_pgt, :casserver_proxy_granting_tickets
+      else
+        drop_table :casserver_pgt
+        drop_table :casserver_tgt
+        drop_table :casserver_st
+        drop_table :casserver_lt
+      end
     end
   end
 end
